getting closer, model values update and notify views just need to add save functionality

+++ src/act/model/ModelEvent.java
@@ -3,10 +3,10 @@ import java.awt.event.ActionEvent;
 
 public class ModelEvent extends ActionEvent {
 	private static final long serialVersionUID = 1L;
-	private int amount;
-	public ModelEvent(Object obj, int id, String message, int amount){
+	private double amount;
+	public ModelEvent(Object obj, int id, String message, double amount){
 		super(obj, id, message);
 		this.amount = amount;
 	}
-	public int getAmount(){return amount;}
+	public double getAmount(){return amount;}
 }
